Issue#95 - Provide update of security provider elements
